task12/statistics_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6  # psycopg2 for Windows
celery==5.3.1
redis==4.6.0
requests==2.31.0
drf-spectacular==0.26.2
django-filter==23.5

black~=23.7
flake8~=6.1
isort~=5.12